requirements.txt
pyserial
paho-mqtt
pytz